Main app refactoring.
