--- src/main/scala/org/ergoplatform/explorer/Configuration.scala
@@ -1,8 +1,0 @@
-package org.ergoplatform.explorer
-
-import org.ergoplatform.explorer.config.ExplorerConfig
-import pureconfig.loadConfigOrThrow
-
-trait Configuration {
-  val cfg: ExplorerConfig = loadConfigOrThrow[ExplorerConfig]
-}
