+++ src/main/scala/org/ergoplatform/explorer/grabber/OffChainGrabberService.scala
@@ -1,7 +1,7 @@
 package org.ergoplatform.explorer.grabber
 
+import cats.effect.IO
 import cats.effect.concurrent.Ref
-import cats.effect.{ContextShift, IO, Timer}
 import com.typesafe.scalalogging.Logger
 import org.ergoplatform.explorer.config.ExplorerConfig
 import org.ergoplatform.explorer.grabber.http.{NodeAddressService, RequestServiceImpl}
@@ -9,36 +9,25 @@ import org.ergoplatform.explorer.grabber.protocol.ApiTransaction
 import org.ergoplatform.explorer.persistence.TransactionsPool
 
 import scala.concurrent.ExecutionContext
+import scala.concurrent.duration.FiniteDuration
 
 /** Unconfirmed transactions pool monitoring service.
   */
 class OffChainGrabberService(txPoolRef: Ref[IO, TransactionsPool], config: ExplorerConfig)
-                            (implicit ec: ExecutionContext) {
+                            (implicit protected val ec: ExecutionContext)
+  extends LoopedIO {
 
-  private implicit val timer: Timer[IO] = IO.timer(ec)
-  private implicit val cs: ContextShift[IO] = IO.contextShift(ec)
+  protected val logger = Logger("off-chain-grabber-service")
 
-  private val logger = Logger("off-chain-grabber-service")
-
-  private val pause = config.grabber.offChainPollDelay
+  protected val pause: FiniteDuration = config.grabber.offChainPollDelay
 
   private val requestService = new RequestServiceImpl[IO]
 
   private val addressServices = NodeAddressService(config.grabber.nodes.head)
 
-  private val syncTask: IO[Unit] = for {
+  protected val task: IO[Unit] = for {
     _ <- IO(logger.info("Starting off-chain monitoring task."))
     txs <- requestService.get[List[ApiTransaction]](addressServices.memPoolUri)
   } yield txPoolRef.update(_ => TransactionsPool.empty.put(txs))
-
-  private def run(io: IO[Unit]): IO[Unit] = io.attempt
-    .flatMap {
-      case Right(_) => IO.unit
-      case Left(f) => IO(logger.error("An error has occurred: ", f))
-    }
-    .flatMap(_ => IO.sleep(pause))
-    .flatMap(_ => IO.suspend(run(io)))
-
-  def start: IO[Unit] = run(syncTask)
 
 }
