--- src/main/scala/org/ergoplatform/explorer/http/Rest.scala
@@ -1,25 +1,0 @@
-package org.ergoplatform.explorer.http
-
-import akka.http.scaladsl.server.Route
-import org.ergoplatform.explorer.Services
-import org.ergoplatform.explorer.http.handlers._
-
-import scala.concurrent.ExecutionContext
-
-trait Rest extends CorsHandler { self: Services =>
-
-  implicit val ec: ExecutionContext
-
-  val routes: Route = corsHandler(handlers.reduce(_ ~ _))
-
-  private def handlers = List(
-    new BlocksHandler(blocksService).route,
-    new TransactionsHandler(txService).route,
-    new AddressesHandler(addressesService, txService).route,
-    new StatsHandler(statsService).route,
-    new ChartsHandler(statsService).route,
-    new InfoHandler(statsService).route,
-    new SearchHandler(blocksService, txService, addressesService, minerService).route
-  )
-
-}
